+++ frontend/app/page.module.css
@@ -1,0 +1,24 @@
+#title {
+  font-family: 'Gelio', sans-serif;
+  font-weight: bold;
+  color: white;
+}
+
+.box {
+  width: 200px;
+  height: 100px;
+  background-image: url(/img/CloudL.jpg);
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: cover;
+}
+
+.box2 {
+  position: absolute;
+  width: 200px;
+  height: 100px;
+  background-image: url(/img/CloudR.png);
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: cover;
+}
