+++ frontend/package.json
@@ -20,6 +20,7 @@
     "@ducanh2912/next-pwa": "^8.2.0",
     "@react-three/drei": "^9.65.3",
     "@react-three/fiber": "^8.12.0",
+    "framer-motion": "^10.12.12",
     "glsl-random": "^0.0.5",
     "next": "^13.4.2",
     "react": "^18.2.0",
