+++ frontend/public/fonts/gelio/Gelio Greek - ReadMe.txt
@@ -1,0 +1,16 @@
+
+GELIO GREEK
+A TrueType Font Family by S. John Ross
+Cumberland Games & Diversions
+www.cumberlandgames.com
+
+When I released Gelio Greek Diner as the "pilot" member of the "Gelio Greek" family, the idea was that I'd trickle out some more styles one at a time. It was a reasonable plan ... but then I remembered I'm really bad at things like "releasing stuff on a schedule." In the spirit of making sure the finished members of the family don't languish unused on my hard drive, I've added four new styles to this release (Gelio Fasolada, Gelio Kleftiko, Gelio Pasteli and Gelio Retsina). Pasteli, Fasolada and Retsina are in many ways the simplest "core" versions of the font: warm and slightly irregular master-glyphs in three weights. Gelio Kleftiko is a "roughly engraved" variant (compatible in both texture and light-angle with other engraved Cumberland fonts). There's more Gelio on the horizon (no bets on when), so keep an eye peeled for future releases.
+
+Gelio is an all-caps font complete with a full set of alphabetical alternates. Mix and match! Gelio is Greek for "laughter," and the family's figures are suited to its name: casual, canted this way and that, sometimes unrelated in height, and strung along a centerline rather than built upward from a baseline. It's randomly playful in a way that required an unexpected amount of planning! Since the font has an awkward, asses-and-elbows nature (and because it's a dual set of interlocking caps) it required more extensive hand-kerning than any other font in the Cumberland library. Now I'm all tired out and craving souvlaki ... but I'm real happy with the font and hope you are, too.
+
+These fonts are each copyright (c) 2011 by S. John Ross. "Cumberland Games & Diversions" and the title of each font are trademarks of S. John Ross. These fonts are free for private use only. Any public or commercial use, or any use by an organization rather than an individual, requires a license; contact the Cumberland Fontworks via email (sjohn@cumberlandgames.com) or find me on the Web for additional contact information. Additional glyphs (foreign alphabet support, customizations, etc) are available by commission.
+
+This ZIP archive may be freely distributed provided none of the contents are altered or removed.
+
+Version 1.0 (Version 1.01 in the case of Gelio Greek Diner)
+
